requirements.txt
certifi==2020.11.8
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==19.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
matplotlib=3.3.3
numpy==1.19.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
statsmodel=0.12.1
Werkzeug==0.16.0
wincertstore==0.2